+++ internal/sanitycheck.php
@@ -62,25 +62,32 @@ function getParameterCountsFromReflection(array $args) : array
     return [$num_required, $num_optional];
 }
 
-// TODO: reuse code?
+/**
+ * This represents an entry in Phan's internal function signatures for a function/method parameter.
+ *
+ * (e.g. `'...args' => 'string|int'`)
+ */
 class PhanParameterInfo
 {
-    /** @var string */
+    /** @var string the parameter name */
     public $name;
+
     /**
-     * @var string
+     * @var string the original spec string, e.g. `'...args'`
      * @suppress PhanWriteOnlyPublicProperty may use this in the future, e.g. for warnings
      */
     public $original_name_spec;
 
-    /** @var bool */
+    /** @var bool is this parameter optional */
     public $is_optional;
-    /** @var bool */
+
+    /** @var bool is this parameter passed by reference */
     public $is_by_reference;
-    /** @var bool */
+
+    /** @var bool is this parameter variadic */
     public $is_variadic;
 
-    /** @var string */
+    /** @var string the union type string */
     public $value;
 
     public function __construct(string $original_name_spec, string $value_spec)