+++ src/Phan/Language/Element/Clazz.php
@@ -30,16 +30,23 @@ use Phan\Language\UnionType;
 use Phan\Library\None;
 use Phan\Library\Option;
 use Phan\Library\Some;
+use Phan\Memoize;
 use Phan\Plugin\ConfigPluginSet;
 
 use Closure;
 
 /**
+ * Clazz represents the information Phan knows about a class, trait, or interface,
+ * the state of Phan populating that information (hydration),
+ * and methods to access that information.
+ *
+ * @see CodeBase for the data structures used for looking up classes or elements of classes (properties, methods, constants, etc)
+ *
  * @phan-file-suppress PhanPartialTypeMismatchArgument
  */
 class Clazz extends AddressableElement
 {
-    use \Phan\Memoize;
+    use Memoize;
     use ClosedScopeElement;
 
     /**
@@ -2524,7 +2531,13 @@ class Clazz extends AddressableElement
 
     public function getMarkupDescription() : string
     {
+        $fqsen = $this->getFQSEN();
         $string = '';
+        $namespace = ltrim($fqsen->getNamespace(), '\\');
+        if ($namespace !== '') {
+            // Render the namespace one line above the class
+            $string .= "namespace $namespace;\n";
+        }
 
         if ($this->isFinal()) {
             $string .= 'final ';
@@ -2542,8 +2555,7 @@ class Clazz extends AddressableElement
             $string .= 'class ';
         }
 
-        // TODO: Also render the namespace?
-        $string .= (string)$this->getFQSEN()->getName();
+        $string .= $fqsen->getName();
         return $string;
     }