+++ src/Phan/Language/Scope/GlobalScope.php
@@ -4,8 +4,9 @@ namespace Phan\Language\Scope;
 use Phan\Language\Element\Variable;
 use Phan\Language\Scope;
 
-// TODO: Warn if __clone() is called, the code is doing something wrong.
-//       the underlying static $global_variable_map isn't cloned, so it's a no-op.
+/**
+ * Represents the global scope (and stores global variables)
+ */
 class GlobalScope extends Scope
 {