+++ src/Phan/Output/Filter/MinimumSeverityFilter.php
@@ -5,10 +5,14 @@ use Phan\Issue;
 use Phan\IssueInstance;
 use Phan\Output\IssueFilterInterface;
 
+/**
+ * MinimumSeverityFilter is a filter that will filter out issues
+ * that are less severe than the provided minimum severity.
+ */
 final class MinimumSeverityFilter implements IssueFilterInterface
 {
 
-    /** @var int */
+    /** @var int the provided minimum severity */
     private $minimum_severity;
 
     /**